+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/SearchSuggestionProvider.kt
@@ -201,9 +201,9 @@ class SearchSuggestionProvider private constructor(
                     // Don't show an autocomplete arrow for the entered text
                     editSuggestion = if (item == text) null else item,
                     icon = icon ?: client.searchEngine?.icon,
-                    // Reducing MAX_VALUE by 2: To allow SearchActionProvider to go above and
-                    // still have one additional spot above available.
-                    score = Int.MAX_VALUE - (index + 2),
+                    // Reducing MAX_VALUE to allow SearchActionProvider and SearchTermSuggestionsProvider
+                    // to go above for which they need additional spots above available.
+                    score = Int.MAX_VALUE - (index + SEARCH_TERMS_MAXIMUM_ALLOWED_SUGGESTIONS_LIMIT + 2),
                     onSuggestionClicked = {
                         searchUseCase.invoke(item)
                         emitSearchSuggestionClickedFact()